+++ src/lexograph/layout/__init__.py
@@ -3,11 +3,13 @@
 from lexograph.layout.linear import linear_layout
 from lexograph.layout.spiral import spiral_layout, tangent_angles
 from lexograph.layout.walk import heading_angles, walk_layout
+from lexograph.layout.walk3d import walk3d_layout
 from lexograph.layout.widths import rendered_widths
 
 __all__ = [
     "linear_layout",
     "walk_layout",
+    "walk3d_layout",
     "heading_angles",
     "spiral_layout",
     "tangent_angles",